requirements.txt
streamlit>=1.28
python-dotenv
requests
beautifulsoup4
openai
msal>=1.24
pandas
openpyxl
